Refactor all the file names into using kebab-case JC-623
