+++ apps/vanilla/src/pages/blackjack/models/index.ts
@@ -1,2 +1,2 @@
-export * from './Subscriber';
-export * from './Publisher';
+export * from './subscriber';
+export * from './publisher';
