+++ apps/vanilla/src/pages/blackjack/models/subscriber.ts
