+++ apps/vanilla/src/pages/blackjack/script/dice-generator.ts
